+++ pleroma_bot/_pleroma.py
@@ -3,10 +3,11 @@ from json.decoder import JSONDecodeError
 
 import re
 import json
+import time
 import shutil
 import requests
 import mimetypes
-from datetime import datetime, timedelta
+from datetime import datetime, timedelta, timezone
 
 # Try to import libmagic
 # if it fails just use mimetypes
@@ -20,6 +21,56 @@ from .i18n import _
 from ._utils import random_string, guess_type
 
 
+def pleroma_api_request(method, url,
+                        params=None, data=None, headers=None, cookies=None,
+                        files=None, auth=None, timeout=None, proxies=None,
+                        hooks=None, allow_redirects=True, stream=None,
+                        verify=None, cert=None, json=None):  # pragma: todo
+    response = requests.request(
+        method=method.upper(),
+        url=url,
+        headers=headers,
+        files=files,
+        data=data or {},
+        json=json,
+        params=params or {},
+        auth=auth,
+        cookies=cookies,
+        hooks=hooks,
+    )
+    if response.status_code == 429:
+        remaining_header = response.headers.get("X-RateLimit-Remaining")
+        limit_header = response.headers.get("X-RateLimit-Limit")
+        reset_header = response.headers.get("X-RateLimit-Reset")
+        if remaining_header and limit_header and reset_header:
+            remaining = int(remaining_header)
+            limit = int(limit_header)
+            ts = datetime.strptime(
+                reset_header,
+                "%Y-%m-%dT%H:%M:%S.%fZ"
+            ).replace(tzinfo=timezone.utc).timestamp()
+            reset_time = datetime.utcfromtimestamp(ts)
+
+            logger.info(_(
+                "Rate limit exceeded. {} out of {} requests remaining until {}"
+                " UTC"
+            ).format(remaining, limit, reset_time))
+
+            delay = (reset_time - datetime.utcnow()).total_seconds() + 2
+            logger.info(_("Sleeping for {}s...").format(round(delay)))
+            time.sleep(delay)
+
+            response = pleroma_api_request(method, url,
+                                           params=params, data=data,
+                                           headers=headers, cookies=cookies,
+                                           files=files, auth=auth, hooks=hooks,
+                                           timeout=timeout, proxies=proxies,
+                                           allow_redirects=allow_redirects,
+                                           stream=stream, verify=verify,
+                                           cert=cert, json=json)
+    return response
+
+
 def get_date_last_pleroma_post(self):
     """Gathers last post from the user in Pleroma and returns the date
     of creation.
@@ -30,7 +81,11 @@ def get_date_last_pleroma_post(self):
         f"{self.pleroma_base_url}/api/v1/accounts/"
         f"{self.pleroma_username}/statuses"
     )
-    response = requests.get(pleroma_posts_url, headers=self.header_pleroma)
+    response = pleroma_api_request(
+        'GET',
+        pleroma_posts_url,
+        headers=self.header_pleroma,
+    )
     if not response.ok:
         response.raise_for_status()
     posts = json.loads(response.text)
@@ -90,8 +145,10 @@ def post_pleroma(
         pleroma_reblog_url = (
             f"{self.pleroma_base_url}/api/v1/statuses/{post_id}/reblog"
         )
-        response = requests.post(
-            pleroma_reblog_url, headers=self.header_pleroma
+        response = pleroma_api_request(
+            'POST',
+            pleroma_reblog_url,
+            headers=self.header_pleroma,
         )
         if not response.ok:
             response.raise_for_status()
@@ -145,9 +202,10 @@ def post_pleroma(
                             "description": alt_text
                         }
                     )
-                response = requests.post(
+                response = pleroma_api_request(
+                    'POST',
                     pleroma_media_url,
-                    data,
+                    data=data,
                     headers=self.header_pleroma,
                     files=files
                 )
@@ -222,8 +280,11 @@ def post_pleroma(
 
     empty = (tweet_text == '' and len(media_ids) == 0 and poll is None)
     if not empty:
-        response = requests.post(
-            pleroma_post_url, data, headers=self.header_pleroma
+        response = pleroma_api_request(
+            'POST',
+            pleroma_post_url,
+            data=data,
+            headers=self.header_pleroma,
         )
         if not response.ok:
             response.raise_for_status()
@@ -324,8 +385,12 @@ def update_pleroma(self):
         )
         files.update({"header": (header_file_name, header, header_mime_type)})
 
-    response = requests.patch(
-        cred_url, data, headers=self.header_pleroma, files=files
+    response = pleroma_api_request(
+        'PATCH',
+        cred_url,
+        data=data,
+        headers=self.header_pleroma,
+        files=files
     )
     try:
         if not response.ok:
@@ -352,8 +417,11 @@ def _pleroma_update_bot_status(self, bot):  # pragma: todo
         f"{self.pleroma_base_url}/api/v1/accounts/update_credentials"
     )
     data = {"bot": str(bot).lower()}
-    response = requests.patch(
-        update_cred_url, headers=self.header_pleroma, data=data
+    response = pleroma_api_request(
+        'PATCH',
+        update_cred_url,
+        data=data,
+        headers=self.header_pleroma,
     )
     if not response.ok:
         response.raise_for_status()
@@ -364,7 +432,11 @@ def _get_pleroma_profile_info(self):  # pragma: todo
         f"{self.pleroma_base_url}/api/v1/accounts/"
         f"{self.pleroma_username}"
     )
-    response = requests.get(profile_url, headers=self.header_pleroma)
+    response = pleroma_api_request(
+        'GET',
+        profile_url,
+        headers=self.header_pleroma,
+    )
     if not response.ok:
         response.raise_for_status()
     profile = response.json()